backend/pyproject.toml
[tool.poetry]
name = "phosphor-backend"
version = "0.1.0"
description = "AI research tool for labs - Lab State Compressor"
authors = ["Phosphor Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.0"
litellm = "^1.55.0"
tiktoken = "^0.8.0"
httpx = "^0.28.0"
pyjwt = {extras = ["crypto"], version = "^2.10.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.24.0"
pytest-env = "^1.1.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
bandit = {extras = ["toml"], version = "^1.8.0"}
pip-audit = "^2.7.0"
httpx = "^0.28.0"
types-redis = "^4.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "S",      # flake8-bandit
    "T20",    # flake8-print
    "SIM",    # flake8-simplify
]
ignore = ["S101"]  # Allow assert in tests

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["celery.*", "litellm.*", "tiktoken.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"

[tool.pytest_env]
DATABASE_URL = "postgresql+asyncpg://phosphor:phosphor@localhost:5432/phosphor_test"
REDIS_URL = "redis://localhost:6379/1"
CLERK_SECRET_KEY = "test_secret_key"
ANTHROPIC_API_KEY = "test_api_key"
ENVIRONMENT = "test"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]  # Skip assert warnings

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__pycache__/*"]
